requirements.txt
Flask==3.1.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.20.1
Pillow==10.4.0
rembg==2.0.67
